+++ model/src/main/java/model/dao/DBConnection.java
@@ -1,4 +1,4 @@
-package model;
+package model.dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
