+++ model/src/main/java/model/dao/DBProperties.java
@@ -1,4 +1,4 @@
-package model;
+package model.dao;
 
 import java.io.IOException;
 import java.io.InputStream;
